requirements.txt
# Adobe Stock Image Processor - Python Dependencies
# 
# Core dependencies for image processing and analysis
# Install with: pip install -r requirements.txt

# FastAPI and web framework dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0

# Database dependencies
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.12.0

# Image processing and computer vision
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# Machine learning and AI (optional - will work without these)
# tensorflow>=2.13.0
# scikit-learn>=1.3.0

# Data processing and analysis
pandas>=2.0.0
openpyxl>=3.1.0

# Report generation and visualization
matplotlib>=3.7.0
seaborn>=0.12.0
jinja2>=3.1.0

# Image hashing and similarity detection
imagehash>=4.3.1

# Progress tracking and user interface
tqdm>=4.65.0

# System monitoring and utilities
psutil>=5.9.0

# Development and testing dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
memory-profiler>=0.61.0